+++ sentry/src/main/java/io/sentry/TraceContext.java
@@ -26,9 +26,25 @@ public final class TraceContext implements JsonUnknown, JsonSerializable {
   private @Nullable Map<String, @NotNull Object> unknown;
 
   TraceContext(@NotNull SentryId traceId, @NotNull String publicKey) {
-    this(traceId, publicKey, null, null, null, null, null, null, null);
+    this(traceId, publicKey, null, null, null, null, null, null);
   }
 
+  TraceContext(
+      @NotNull SentryId traceId,
+      @NotNull String publicKey,
+      @Nullable String release,
+      @Nullable String environment,
+      @Nullable String userId,
+      @Nullable String transaction,
+      @Nullable String sampleRate,
+      @Nullable String sampled) {
+    this(traceId, publicKey, release, environment, userId, null, transaction, sampleRate, sampled);
+  }
+
+  /**
+   * @deprecated segment has no effect and will be removed in the next major update.
+   */
+  @Deprecated
   TraceContext(
       @NotNull SentryId traceId,
       @NotNull String publicKey,